Move about to Home package since it's not petlist related
