+++ samples/star/src/main/kotlin/com/slack/circuit/star/home/About.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.slack.circuit.star.petlist
+package com.slack.circuit.star.home
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.layout.Box
